feat: handle closed connection
